+++ data/city_sample.csv
@@ -1,1 +1,2 @@
 first_name,last_name,email_address,home_phone,street,city,state,zipcode
+Aya,Fuller,jtex@jumpstartlab.com,778.232.7000,2-1325 Barclay Street,Vancouver,Bc,90210
