+++ test/queue_test.rb
@@ -17,4 +17,15 @@ class ListTest < Minitest::Test
     assert_equal 35, er.print_sorted(:last_name).count
   end
 
+  def test_queue_can_be_cleared
+    er = EventReporter.new('./data/full_event_attendees.csv')
+    list = List.new
+
+    er.find_attendees(:first_name, "Rachel")
+    refute er.list.empty?
+
+    list.clear_queue
+    assert_equal 0, list.queue_count
+  end
+
 end
